peliculas/requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.0
django-crispy-forms==1.11.2
django-environ==0.4.5
django-filter==2.2.0
django-heroku==0.3.1
django-multiupload==0.5.2
gunicorn==20.1.0
Pillow==8.1.2
psycopg2==2.8.6
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
